moved and renamen InitialDataLoader and InitialDataLoaderConfig
