+++ dsf-fhir/dsf-fhir-server/src/main/java/org/highmed/dsf/fhir/service/InitialDataLoader.java
@@ -1,0 +1,8 @@
+package org.highmed.dsf.fhir.service;
+
+import org.hl7.fhir.r4.model.Bundle;
+
+public interface InitialDataLoader
+{
+	void load(Bundle bundle);
+}
