+++ dsf-fhir/dsf-fhir-server/src/main/java/org/highmed/dsf/fhir/service/InitialDataLoaderImpl.java
@@ -1,4 +1,4 @@
-package org.highmed.dsf.fhir.init;
+package org.highmed.dsf.fhir.service;
 
 import java.util.Objects;
 
@@ -12,14 +12,14 @@ import org.springframework.beans.factory.InitializingBean;
 
 import ca.uhn.fhir.context.FhirContext;
 
-public class InitialDataLoader implements InitializingBean
+public class InitialDataLoaderImpl implements InitialDataLoader, InitializingBean
 {
-	private static final Logger logger = LoggerFactory.getLogger(InitialDataLoader.class);
+	private static final Logger logger = LoggerFactory.getLogger(InitialDataLoaderImpl.class);
 
 	private final CommandFactory commandFactory;
 	private final FhirContext fhirContext;
 
-	public InitialDataLoader(CommandFactory commandFactory, FhirContext fhirContext)
+	public InitialDataLoaderImpl(CommandFactory commandFactory, FhirContext fhirContext)
 	{
 		this.commandFactory = commandFactory;
 		this.fhirContext = fhirContext;
@@ -32,6 +32,7 @@ public class InitialDataLoader implements InitializingBean
 		Objects.requireNonNull(fhirContext, "fhirContext");
 	}
 
+	@Override
 	public void load(Bundle bundle)
 	{
 		if (bundle == null)
@@ -41,14 +42,13 @@ public class InitialDataLoader implements InitializingBean
 		}
 
 		CommandList commands = commandFactory.createCommands(bundle);
-		logger.info("Executing command list for bundle with {} entries", bundle.getEntry().size());
+		logger.debug("Executing command list for bundle with {} entries", bundle.getEntry().size());
 		Bundle result = commands.execute();
 		result.getEntry().forEach(this::logResult);
 	}
 
 	private void logResult(BundleEntryComponent entry)
 	{
-		logger.info("Initial data loaded: {} - {}", entry.getResponse().getStatus(),
-				entry.getResponse().getLocation());
+		logger.info("{} {}", entry.getResponse().getLocation(), entry.getResponse().getStatus());
 	}
 }
